+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Allow last focused non selectable terminal to be focused when universal selection is turned on.
 - Add `size` option for split layouts, accepting both numeric and percentage values.
+- Make floating terminals resize
 
 ## [0.1.2] - 2025-07-24